forensics_service/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
numpy==2.2.3
opencv-python==4.11.0.86
scikit-image==0.25.2
Pillow==11.1.0
exifread==3.0.0
hachoir==3.3.0
librosa==0.10.2.post1
torchaudio==2.6.0
torch==2.6.0
mediapipe==0.10.21
reportlab==4.3.1
python-dateutil==2.9.0.post0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
pymongo==4.12.1
